### 2021-07-23 #3531 @vinisalazar
Hi @danielskatz and @jmschrei. Thank you for the prompt handling of the paper. I will make the changes as soon as possible and report back here. Have a nice weekend!

Best,
Vini
