### 2021-07-26 202fd7e
Fix references and affiliation
