+++ paper/paper.bib
@@ -15,7 +15,8 @@
   number    = {W1},
   pages     = {W537--W544},
   year      = {2018},
-  publisher = {Oxford University Press}
+  publisher = {Oxford University Press},
+  doi       = {10.1093/nar/gkaa554}
 }
 @article{de2012provenance,
   title     = {A provenance-based adaptive scheduling heuristic for parallel scientific workflows in clouds},
@@ -25,7 +26,8 @@
   number    = {3},
   pages     = {521--552},
   year      = {2012},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi       = {10.1007/s10723-012-9227-2}
 }
 @book{rse-py,
   author    = {Damien Irving and Kate Hertweck and Luke Johnston and Joel Ostblom and Charlotte Wickham and Greg Wilson},
@@ -79,7 +81,8 @@
   title   = {{Why and Where: A Characterization of Data Provenance}},
   url     = {http://homepages.inf.ed.ac.uk/opb/papers/ICDT2001.pdf},
   volume  = {1973},
-  year    = {2001}
+  year    = {2001},
+  doi     = {10.1007/3-540-44503-x_20}
 }
 @article{camacho2009blast,
   author    = {Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L},
@@ -228,15 +231,16 @@
   volume    = {35},
   year      = {2017}
 }
-@book{Groth2013,
-  author    = {Groth, Paul and Moreau, Luc},
-  doi       = {10.2200/S00528ED1V01Y201308WEB007},
-  editor    = {Hendler, James and Ding, Ying},
-  isbn      = {9781627052214},
-  pages     = {131},
-  publisher = {Morgan {\&} Claypool},
-  title     = {{An Introduction to PROV}},
-  year      = {2013}
+@article{Groth2013,
+  title     = {Provenance: an introduction to PROV},
+  author    = {Moreau, Luc and Groth, Paul},
+  journal   = {Synthesis lectures on the semantic web: theory and technology},
+  volume    = {3},
+  number    = {4},
+  pages     = {1--129},
+  year      = {2013},
+  publisher = {Morgan \& Claypool Publishers},
+  doi       = {10.2200/S00528ED1V01Y201308WBE007}
 }
 @article{Hyatt2010,
   abstract = {BACKGROUND: The quality of automated gene prediction in microbial organisms has improved steadily over the past decade, but there is still room for improvement. Increasing the number of correct identifications, both of genes and of the translation initiation sites for each gene, and reducing the overall number of false positives, are all desirable goals.$\backslash$n$\backslash$nRESULTS: With our years of experience in manually curating genomes for the Joint Genome Institute, we developed a new gene prediction algorithm called Prodigal (PROkaryotic DYnamic programming Gene-finding ALgorithm). With Prodigal, we focused specifically on the three goals of improved gene structure prediction, improved translation initiation site recognition, and reduced false positives. We compared the results of Prodigal to existing gene-finding methods to demonstrate that it met each of these objectives.$\backslash$n$\backslash$nCONCLUSION: We built a fast, lightweight, open source gene prediction program called Prodigal http://compbio.ornl.gov/prodigal/. Prodigal achieved good results compared to existing methods, and we believe it will be a valuable asset to automated microbial annotation pipelines.},
