### 2021-07-26 19523da
Replace classes figure
  - Replace .svg with .png figure so paper compiles correctly.
  - Change figure path in paper.md.
