### 2021-08-16 #3622 @whedon
[REVIEW]: BioProv - A provenance library for bioinformatics workflows
**Submitting author:** @vinisalazar (<a href="http://orcid.org/0000-0002-8362-3195">Vinícius Salazar</a>)
**Repository:** <a href="https://github.com/vinisalazar/BioProv/" target ="_blank">https://github.com/vinisalazar/BioProv/</a>
**Version:** v0.1.24
