### 2021-09-13 #3622 @Zethson
Dear @vinisalazar,

thank you very much for your submission. I had quite some fun reading it and think that your approach to the outlined issue is very interesting. I come more from a classical bioinformatics workflow background and found your take with a "Project" centric view a novel view on the problem.

I have a couple of points/questions that I would be very happy to see addressed:

1. What is your reasoning behind making the `Project` be executable with BioProv? I personally think that traditional workflow engines such as Nextflow or Snakemake already do this but with many many more features (inherent parallelization, container support, cloud support, [...]).
2. BioProv is supposed to improve reproducibility, but the lacking container support and computational portability make this difficult. The workflow engines also track execution time, execution place etc. What novelty does the W3C-Prov report introduce to improve reproducibility that the workflow engines do not yet cover?
3. I personally think that the USP of BioProv should be the generation of W3C-Prov compatible reports. What do you think about introducing a Nextflow DSL2 module or Snakemake module to facilitate the generation of such reports for all pipelines automatically?

Thank you very much. Really looking forward to your answers.
- Lukas
