### 2021-09-13 #3622 @vinisalazar
Dear @Zethson,

Thank you for your comments. Here are our responses:

1. The main rationale behind this is to be able to consume processes' start and end time, standard output and error, and other relevant information from within the main BioProv process. We aimed to provide a simple framework that does this using Python standard library modules. I agree that there are excellent open-source workflow engines and I use them daily, but we wanted BioProv to be able to run programs natively (and not to depend on third-party tools), even with limited functionality in comparison to more established frameworks. We have added these arguments to the text to make it clearer.

2. The main novelty of the BioProv report in comparison to other engines is representing a workflow instance (a project) as a W3C-PROV compatible object, with support for both prospective (*i.e.* information previous to a project's execution) and retrospective provenance. The W3C-PROV recommendation is a W3C standard that has extensive technical reasoning behind it. Our package aims to introduce how this standard can be utilized and disseminated in bioinformatics, to improve the integration of bioinformatic analyses with other knowledge bases.

3. I would be delighted to see a Nextflow or Snakemake plugin offering generation of W3C-PROV reports. I believe that the latter is more feasible since BioProv and Snakemake are both written in Python. However, we did want to introduce a standalone library as a proof of concept on how to fit bioinformatics workflows into the existing W3C-PROV framework. We hope to attract contributors to our open source code to potentially develop these plugins.

Please let us know if there are any other questions.

Best wishes,

Vinícius, on behalf of the BioProv team
