### 2021-09-15 #3622 @Zethson
Dear @vinisalazar,

thank you very much for your answers. I agree with your statements. Nevertheless, I still think that that a plugin for a workflow language could have good impact and hope that you will revisit this post publication (if no one from the OS community volunteers).

I found a typo in line 170: should be "containing"

@jmschrei I consider the paper to be well written with a novel view on the outlined issue. The paper is accompanied with useful tutorials and examples enabling a quick start.

I would like to recommend it for publication.

Cheers
- Lukas
