### 2021-09-15 144e4ac
Fix typo
  Review from openjournals/joss-reviews#3622
