### 2021-10-05 vinisalazar/BioProv#51 @maximtrp
Package requirements
Hello!

While reviewing your package, I have noticed that the requirements are not clearly stated anywhere (just several packages mentioned here and there). Please add such a section to your documentation and README file. You can just copy-paste the requirements from your [setup.py](https://github.com/vinisalazar/BioProv/blob/master/setup.py) file.

https://github.com/openjournals/joss-reviews/issues/3622
