### 2021-10-05 7cff99c
Update environment.yaml
  - Make environment file more concise, use mostly dependencies from setup.py
  - Add dependencies from environment.yaml to README
