+++ environment.yml
@@ -4,82 +4,16 @@ channels:
   - bioconda
   - defaults
 dependencies:
-  - attrs
-  - bioprov
   - biopython
-  - black
-  - brotlipy
-  - ca-certificates
-  - certifi
-  - cffi
-  - chardet
   - coolname
-  - coverage
   - coveralls
-  - cryptography
   - dataclasses
-  - decorator
-  - docopt
-  - html5lib
-  - icu
-  - idna
-  - iniconfig
-  - isodate
-  - jupyterlab
-  - keepalive
-  - libblas
-  - libcblas
-  - libcxx
-  - libffi
-  - libgfortran
-  - libgfortran5
-  - libiconv
-  - liblapack
-  - libopenblas
-  - libxml2
-  - libxslt
-  - llvm-openmp
-  - lxml
-  - more-itertools
-  - ncurses
-  - networkx
-  - numpy
-  - openssl
-  - packaging
   - pandas
-  - pip
-  - pluggy
   - prodigal
   - prov
   - provstore-api
-  - py
-  - pycparser
   - pydot
-  - pyopenssl
-  - pyparsing
-  - pysocks
   - pytest
   - pytest-cov
-  - python
-  - python-dateutil
-  - python_abi
-  - pytz
-  - pyyaml
-  - rdflib
-  - readline
-  - requests
-  - setuptools
-  - six
-  - sparqlwrapper
-  - sqlite
-  - tinydb
-  - tk
-  - toml
   - tqdm
-  - tzdata
-  - urllib3
-  - webencodings
-  - wheel
-  - xz
-  - yaml
-  - zlib
+  - tinydb
