+++ CHANGELOG.md
@@ -1,184 +1,184 @@
 ## Changelog
 
 ### To-do
-* Fix documentation issues [ ]
-    * Add subpackage docstrings [ ]
-* Create methods for Sample and Project
-    * .describe [ ]
-    * .write_paths_to_file, .copy_files_to_dir(), .link_files_to_dir() [ ]
-    * .total_duration [ ]
-* Add logger calls when saving to JSON and uploading to ProvStore [ ]
+* [ ] Fix documentation issues
+    * [ ] Add subpackage docstrings
+* [ ] Create methods for Sample and Pro
+    * [ ] .describe
+    * [ ] .write_paths_to_file, .copy_files_to_dir(), .link_files_to_dir()
+    * [ ] .total_duration
+* [ ] Add logger calls when saving to JSON and uploading to ProvStore
 
 ### v0.1.24
-* State requirements more clearly [x]
-* Improve error handling when getting `Program` attributes [x]
-* Add `__doc__` variable to modules to build documentation [x]
-* Fix `File.size` property [x]
-* Add `File.raw_size` property [x]
+* [x] State requirements more clearly
+* [x] Improve error handling when getting `Program` attributes
+* [x] Add `__doc__` variable to modules to build documentation
+* [x] Fix `File.size` property
+* [x] Add `File.raw_size` property
 
 ### v0.1.23
-* Patch PresetProgram SeqFile addition feature [x]
-* Fix SeqFile deserializer [x]
-* Add import_records arg to `bp.load_project()` [x]
-* Improve reserved aminoacid characters [x]
-* Add `SeqFile.max_seq` and `.min_seq` properties [x]
-* Patch `Project` deserializer to improve BioProvDocument creation [x]
-* Make shorter Environment hashes [x]
-* Improve Project `__repr__` [x]
+* [x] Patch PresetProgram SeqFile addition feature
+* [x] Fix SeqFile deserializer
+* [x] Add import_records arg to `bp.load_project()`
+* [x] Improve reserved aminoacid characters
+* [x] Add `SeqFile.max_seq` and `.min_seq` properties
+* [x] Patch `Project` deserializer to improve BioProvDocument creation
+* [x] Make shorter Environment hashes
+* [x] Improve Project `__repr__`
 
 ### v0.1.22
-* Simplify `bp.load_project()` function [x]
-* Fix user and env PROV relationships [x]
-* Add Run attributes as Program class properties (will consume from the last run) [x]
-* Add sequence dunder methods for Project class [x]
-* Improve run methods [x]
-* Allow specifying sequence format to PresetProgram outputs [x]
-* Update env file to Py 3.9 [x]
+* [x] Simplify `bp.load_project()` function
+* [x] Fix user and env PROV relationships
+* [x] Add Run attributes as Program class properties (will consume from the last run)
+* [x] Add sequence dunder methods for Project class
+* [x] Improve run methods
+* [x] Allow specifying sequence format to PresetProgram outputs
+* [x] Update env file to Py 3.9
 
 ### v0.1.21
-* Add _config argument to `bp.load_project()` (this is a temporary fix) [x]
-* Add db property to Config class (prevents bug when setting DB path) [x]
-* Improve a few docstrings [x]
-* Refactor sha1 as sha256 [x]   
-* Add add and radd dunder methods for Directory and File [x]  
-* Package workflows in single module [x]
-* Add FastTree PresetProgram [x]
+* [x] Add _config argument to `bp.load_project()` (this is a temporary fix)
+* [x] Add db property to Config class (prevents bug when setting DB path)
+* [x] Improve a few docstrings
+* [x] Refactor sha1 as sha256   
+* [x] Add add and radd dunder methods for Directory and File  
[…]
