### 2021-10-05 vinisalazar/BioProv#51 @vinisalazar
Hi @maximtrp,

thank you for the issue. I have opened a PR to merge the latest version. It makes several improvements to the docs, including what you requested. I will merge it shortly.

Best,
Vini

Edit: I closed this upon the merge, but please feel free to reopen if needed.
