### 2021-11-02 #3622 @vinisalazar
@jmschrei thank you for the message, and @maximtrp thank you for the review.

I have addressed and closed the issues raised. Please let me know if there are any pendencies.

Best,
Vinícius
