### 2021-11-04 #3622 @vinisalazar
Hello @jmschrei, thank you for conducting the editing and review process. Is it okay if I upload the latest tag, [`v0.1.24`](https://github.com/vinisalazar/BioProv/releases/tag/v0.1.24)? That's the current version that's available on PyPI, with all the changes requested from the reviews.
