### 2021-11-05 #3622 @vinisalazar
Apologies, I hadn't realised that had already been done by my CI setup. Here it is: https://doi.org/10.5281/zenodo.5550568

This version contains all changes requested by the reviewers. Please note that the current tag (`v0.1.24`) is different from the originally submitted version (`v0.1.23`).

Please let me know if anything else is needed.

Best wishes,
Vinícius
